Refactor the build system for Linux and Mac OS X
This commit changes the build process for Linux and Mac OS X. It no
longer co-mingles the build of libjs1 with the build of Event Store.
Instead, when building Event Store, we assume a pre-built version of
libjs1.[so|dylib] is available in src/libs/x64/<current distro/version>.

The current distro and version is detected using a modified version of a
script by Boundary (Apache license, see scripts/build-js1/detect-system.sh).
It can also be overriden as the final parameter to the build script.

Pre-build versions of libjs1 are included for:
- Amazon Linux 2015.03
- Ubuntu Linux 14.04 (LTS)
- Mac OS X Yosemite (10.10)

Changes are as follows:

- Remove support for specifying platform (x64 vs x86) - only x64 is
  supported for Event Store.
- Add libjs1 built for Amazon Linux 2015.03
- Add script for building JS1 on different Linux distros
- Add js1 build script for MacOS X
- Remove werror flag setting from build.sh - this is now handled by the
  build scripts for libjs1 only
- Remove actions from the build script for Event Store. The only
  remaining action is "quick" and there is no parameter for this
- Rename libs/x64/ubuntu-trusty -> ubuntu-14.04
- Rename libs/x64/mac -> osx-10.10