Ensure that the AllReader respects IndexCommitter
This is an experimental commit in order to try and reduce the occurence
of a cluster fault which cannot be reproduced. When using the optimized
AllReader path, it may be necessary to gate on
_indexCommitter.LastCommitPosition in order to ensure that events
written past the currently replicated point are not returned. Similarly
when reading backwards this may also be necessary.

Needs a reproducing integration test to commit to the main product; this
is currently for experiementation only and should NOT be considered a
stable commit.